+++ packages/auth-better/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @theoven/auth-better
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @theoven/core@0.6.1
+  - @theoven/auth@0.6.1
+
 ## 0.6.0
 
 ### Patch Changes