+++ packages/auth-mongo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @theoven/auth-mongo
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @theoven/core@0.6.1
+  - @theoven/auth@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes