+++ packages/create-theoven/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-theoven
 
+## 0.6.1
+
+### Patch Changes
+
+- @theoven/cli@0.6.1
+
 ## 0.6.0
 
 ### Patch Changes