+++ packages/db-drizzle/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @theoven/db-drizzle
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @theoven/core@0.6.1
+  - @theoven/db@0.6.1
+
 ## 0.6.0
 
 ### Patch Changes