+++ packages/storage-bunny/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @theoven/storage-bunny
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @theoven/core@0.6.1
+  - @theoven/storage@0.6.1
+
 ## 0.6.0
 
 ### Patch Changes