+++ packages/vector-pg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @theoven/vector-pg
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @theoven/core@0.6.1
+  - @theoven/vector@0.6.1
+
 ## 0.6.0
 
 ### Minor Changes